+++ src/enums.ts
@@ -86,7 +86,7 @@ export enum TrendingType {
 };
 
 // Both filter option property names as well as demo app gradient config keys
-export const DATA_SOURCES_FOR_FILTER_OPTIONS = [
+export const FILTER_OPTION_DATA_SOURCES = [
     ...Object.values(TagTootsCacheKey),  // TODO: these are really the wrong cache keys for the use case but it's consistent w/demo app for now
     ScoreName.FAVOURITED_ACCOUNTS,
 ] as const;