+++ src/types.ts
@@ -10,7 +10,7 @@ import NumericFilter, { NumericFilterArgs } from './filters/numeric_filter';
 import Scorer from './scorer/scorer';
 import TagList from './api/tag_list';
 import Toot, { SerializableToot } from './api/objects/toot';
-import { CacheKey, NonScoreWeightName, ScoreName, TagTootsCacheKey, DATA_SOURCES_FOR_FILTER_OPTIONS } from './enums';
+import { CacheKey, NonScoreWeightName, ScoreName, TagTootsCacheKey, FILTER_OPTION_DATA_SOURCES as FILTER_OPTION_DATA_SOURCES } from './enums';
 
 // Records
 export type AccountNames = Record<mastodon.v1.Account["acct"], Account>;
@@ -38,8 +38,8 @@ export type TootNumberProp = KeysOfValueType<Toot, number>;
 
 
 // Filters
-export type FilterOptionDataSource = typeof DATA_SOURCES_FOR_FILTER_OPTIONS[number];
-export type FilterOptionUserData = {[key in FilterOptionDataSource]?: number};
+export type FilterOptionDataSource = typeof FILTER_OPTION_DATA_SOURCES[number];
+type FilterOptionUserData = {[key in FilterOptionDataSource]?: number};
 export interface BooleanFilterOption extends FilterOptionUserData, ObjWithTootCount {};
 
 export type BooleanFilters = Record<BooleanFilterName, BooleanFilter>;